backend/pyproject.toml
[tool.poetry]
name = "tummytap"
version = "0.1.0"
description = "Dream project food business idea"
authors = ["Sandeepkd1824 <sandeepkd1824@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = ">=5.2,<6.0"
djangorestframework = ">=3.16,<4.0"
djoser = ">=2.3,<3.0"
django-cors-headers = ">=4.7.0,<5.0.0"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"
django-filter = "^25.1"
python-decouple = "^3.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
